secure-chat-pro/web_chat/requirements.txt

flask>=3.0.0
flask-socketio>=5.3.6
python-socketio[client]>=5.10.0
eventlet>=0.36.1
